add AuxIntegrandsExt
