+++ Project.toml
@@ -9,7 +9,6 @@ BaryRational = "91aaffc3-5777-4842-85b7-5d3d5d6a3494"
 EquiBaryInterp = "c5458111-f826-47c0-abe9-f36aa00852ec"
 HChebInterp = "78faba9b-a54b-441f-8118-62407cbe4e59"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
@@ -19,7 +18,6 @@ AutoBZCore = "0.1.6"
 BaryRational = "0.1"
 EquiBaryInterp = "0.1"
 HChebInterp = "0.1"
-Polyhedra = "0.6,0.7"
 Reexport = "1"
 Requires = "1"
 StaticArrays = "1"
@@ -27,12 +25,15 @@ SymmetryReduceBZ = "0.1"
 julia = "1.6"
 
 [extensions]
+AuxIntegrandsExt = "AuxQuad"
 BrillouinPlotlyJSExt = ["Brillouin", "PlotlyJS"]
-SymmetryReduceBZExt = "SymmetryReduceBZ"
+SymmetryReduceBZExt = ["SymmetryReduceBZ", "Polyhedra"]
 
 [extras]
+AuxQuad = "b59dd6d0-509b-4777-bfd1-cd8479d7b9e5"
 Brillouin = "23470ee3-d0df-4052-8b1a-8cbd6363e7f0"
 PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
+Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
 SymmetryReduceBZ = "49a35663-c880-4242-bebb-1ec8c0fa8046"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
@@ -40,6 +41,8 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 test = ["SymmetryReduceBZ", "Test"]
 
 [weakdeps]
+AuxQuad = "b59dd6d0-509b-4777-bfd1-cd8479d7b9e5"
 Brillouin = "23470ee3-d0df-4052-8b1a-8cbd6363e7f0"
 PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
+Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
 SymmetryReduceBZ = "49a35663-c880-4242-bebb-1ec8c0fa8046"
